+++ src/blinker/blinker.txt
@@ -1,0 +1,60 @@
+#include <Arduino.h>
+#include "blinker.h"
+
+#pragma once
+
+
+// DESC:     updates the blinker states in the drive struct
+void update_drive_blinkers(driveState &drive)
+{
+    if (digitalRead(TURN_LEFT_BUTTON))
+    {
+        drive.blink_left = true;
+    }
+    else
+    {
+        drive.blink_left = false;
+    }
+
+    if (digitalRead(TURN_RIGHT_BUTTON))
+    {
+        drive.blink_right = true;
+    }
+    else
+    {
+        drive.blink_right = false;
+    }
+};
+
+void handelBlinkerLeft()
+{
+    blinkers.blink_left != blinkers.blink_left;
+    digitalWrite(BLINKER_LEFT, blinkers.blink_left);
+}
+
+void handelBlinkerRight()
+{
+    blinkers.blink_right != blinkers.blink_right;
+    digitalWrite(BLINKER_RIGHT, blinkers.blink_right);
+}
+
+void flashBlinkers(driveState &drive)
+{
+    // check and update Timers
+    if (drive.blink_left)
+    {
+        blinkLeftTimer.start();
+    }
+    else
+    {
+        blinkLeftTimer.stop();
+    }
+    if (drive.blink_right)
+    {
+        blinkRightTimer.start();
+    }
+    else
+    {
+        blinkRightTimer.stop();
+    }
+}
